Add idstatus. Fix envs. Remove ValEnv. Pats in vb..
Changes were necessary to pave way for handling constructors properly.
Environment handling is more uniform, removing the need for ValEnv since
everywhere now deals with adding to the environment by creating new
nested environments and popping as necessary to preserve scope rather
than passing ValEnvs around.

It's not clear which approach is better, but I need to have a uniform representation.
Wondering if I should switch to using logging of HaMLet directly. It
seems like many of the issues I thought I might have had with that
approach can be mitigated now that I've solved them in my custom AM and
have a much clearer understanding of SML '97.

Added pattern matching to valbinds, but it looks ugly. Not sure the
"right" way to visualize it or if the matching value should go in the
focus or the continuation. Directly affects visualization.